cleaning up code, refactoring location of Mutect.scala, adding in licenses
