+++ avocado-core/src/test/scala/org/bdgenomics/avocado/genotyping/MutectSuite.scala
@@ -1,12 +1,27 @@
-package org.bdgenomics.avocado.algorithms.mutect
+/*
+ * Licensed to Big Data Genomics (BDG) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The BDG licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.bdgenomics.avocado.genotyping
 
 import org.bdgenomics.adam.models.ReferencePosition
 import org.bdgenomics.avocado.models.AlleleObservation
 import org.scalatest.FunSuite
 
-/**
- * Created by john on 5/17/15.
- */
 class MutectSuite extends FunSuite {
   val pos = ReferencePosition("ctg", 101l)
   val mt = new Mutect()
